requirements.txt
pyTelegramBotAPI==4.4.0
pandas==1.4.1
PyGithub==1.55
